Fix bug relating to use of relative paths in CACHE
The CACHE object is used to remember whether a path was a directory or not
across glob() calls. Previously, this cache was keyed by the relative path
of the file (i.e. the basename).

This can result in missing results (not descending into a directory because
it was incorrectly thought to be a file), or an exception (attempting to
descend into a file because it was thought to be a directory).

Now, the CACHE object is instead keyed by the full path, instead of the
relative path.